Add ES6 modules for different funtionalities
